reply table 생성

+++ sb2/src/main/java/org/zerock/sb2/board/dto/BoardReadDTO.java
@@ -9,13 +9,13 @@ public class BoardReadDTO {
     private Long bno;
     private String title,content, writer;
     private boolean delFlag;
-    private int viewCnt;
+    private long viewCnt;
     private LocalDateTime regDate, modDate;
    
    
    //bno,title, content,  writer,  delFlag,  viewCnt,  regDate, modDate
    
-    public BoardReadDTO(Long bno, String title, String content, String writer, boolean delFlag, int viewCnt,
+    public BoardReadDTO(Long bno, String title, String content, String writer, boolean delFlag, long viewCnt,
             LocalDateTime regDate, LocalDateTime modDate) {
         this.bno = bno;
         this.title = title;
